% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeData.R
\name{associationTest}
\alias{associationTest}
\title{Test for an association between two survey questions}
\usage{
associationTest(x, id1, id2)
}
\arguments{
\item{x}{A list where each element represents a single question}

\item{id1}{Index of the first question}

\item{id2}{Index of the second question}
}
\value{
P-value for the test of association
}
\description{
Test for an association between two survey questions
}

