% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean & prepare data obtained from the Typeform API}
\usage{
cleanData(x)
}
\arguments{
\item{x}{Survey data from the typeform api (via ifixitR::getTypeformData(name))}
}
\value{
A list where each element corresponds to survey data for a single question
}
\description{
Clean & prepare data obtained from the Typeform API
}

