% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{compressData}
\alias{compressData}
\title{Compress questions (to remove duplicate data)}
\usage{
compressData(x)
}
\arguments{
\item{x}{Results from one typeform question (data.frame)}
}
\value{
Compressed data in list form
}
\description{
Compress questions (to remove duplicate data)
}

