% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{convertValues}
\alias{convertValues}
\title{Convert response values to the appropriate type (based on question variant)}
\usage{
convertValues(x)
}
\arguments{
\item{x}{Results from one typeform question (data.frame)}
}
\value{
Original data.frame with the freshly converted `value` column
}
\description{
Convert response values to the appropriate type (based on question variant)
}

