% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeData.R
\name{getAssociations}
\alias{getAssociations}
\title{Test associations between all survey questions}
\usage{
getAssociations(data)
}
\arguments{
\item{data}{A data.frame returned from typeformR::cleanData()}
}
\value{
A list that includes:
          data -- data.frame with:  {question1, question1, pvalue}
          plot -- association heatmap
}
\description{
Test associations between all survey questions
}

