% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{getData}
\alias{getData}
\title{Get survey data from Typeform (using the API and rtypeform R package)}
\usage{
getData(api_key = NULL, survey_id = NULL)
}
\arguments{
\item{name}{Survey name}
}
\value{
list containing: {stats, questions, responses, response}
}
\description{
Get survey data from Typeform (using the API and rtypeform R package)
}

