% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeData.R
\name{mergeQuestions}
\alias{mergeQuestions}
\title{Merge user responses for three different questions}
\usage{
mergeQuestions(x, ids)
}
\arguments{
\item{x}{List of data.frames, each with clean results for a single question}

\item{ids}{Vector of question indices}
}
\value{
A list containing {data, type, question}
          data -- data.frame with the merged data: {userid, value1, value2}
}
\description{
Merge user responses for three different questions
}

