% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{parseID}
\alias{parseID}
\title{Parse a Typeform Question ID into its individual components}
\usage{
parseID(x)
}
\arguments{
\item{x}{Typeform question ID (as a character)}
}
\value{
list containing: {type, field_id, choice}
}
\description{
Parse a Typeform Question ID into its individual components
}

