% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeData.R
\name{plotTwoQuestions}
\alias{plotTwoQuestions}
\title{Create a bi-variate plot that describes the results of two questions}
\usage{
plotTwoQuestions(x, id1, id2, levels1 = NULL, levels2 = NULL,
  plot_style = "bar")
}
\arguments{
\item{x}{List of data.frames, each with clean results for a single question}

\item{id1}{Index of the first question}

\item{id2}{Index of the second question}

\item{levels1}{Ordered factor levels for the first question}

\item{levels2}{Ordered factor levels for the second question}

\item{plot_style}{Plot style ("bar", "grid")}
}
\value{
A plot
}
\description{
Create a bi-variate plot that describes the results of two questions
}
\details{
Question type will dictate the plot that is produced:
\itemize{
 \item{\bold{dropdown/list vs. dropdown/list} -- faceted bar graph}
 }
}

