% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{reorderQuestions}
\alias{reorderQuestions}
\title{Put a list of question-responses in the order they were presented}
\usage{
reorderQuestions(x, questions)
}
\arguments{
\item{x}{List with typeform survey results (one question/element)}

\item{questions}{data.frame with questions (i.e., ifixitR::getTypeformData()[['questions']])}
}
\value{
Input list in the correct order
}
\description{
Put a list of question-responses in the order they were presented
}

