% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run a Shiny App included within this package}
\usage{
runShinyApp(myapp)
}
\arguments{
\item{myapp}{The name of the app you want to run.
Submit `runShinyApp()` to see all available apps.}
}
\description{
Run a Shiny App included within this package
}

