% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeData.R
\name{summarizeData}
\alias{summarizeData}
\title{Compute descriptive statistics for a list of typeform question-responses}
\usage{
summarizeData(x)
}
\arguments{
\item{x}{List of data.frames, each with clean results for a single question}
}
\value{
A list where each element is a summary of the correponding question's results
}
\description{
Compute descriptive statistics for a list of typeform question-responses
}
\details{
Text fields are combined into a single data.frame as the last element (called "Text")
}

