% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeData.R
\name{summarizeQuestion}
\alias{summarizeQuestion}
\title{Compute descriptive statistics for responses to a typeform question}
\usage{
summarizeQuestion(x)
}
\arguments{
\item{x}{Data.frame with clean typeform survey results (one question only)}
}
\value{
A data.frame where each element is a summary of the correponding question's results
}
\description{
Compute descriptive statistics for responses to a typeform question
}
\details{
Statistics are computed based on the type of question that is being asked.
}

